==> scATAC-seq/4-clustering/4-level_4/PC/03-peakcalling_level3.Rmd <==
---
title: "Calling Peaks"
author: "Paula Soler-Vila"
output: 
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
---
  
```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE, out.width = "100%", fig.align = "center",
  message = FALSE, warning = FALSE
)
options(width = 1200)
```

# Introduction
Looking at the scATAC-seq coverage plots, it's possible to vislumbrate that, as a consequence of the merging process, some peaks show a considerable increase in size. This can create a problem for certain analysis such as motif enrichment analysis. In order to solve it, we decided to do a new round of peak calling using MACS2 for each group of cells annotated at level 1, helping to identify specifics peaks for rarer populations.  

# Pre-processing

## Load packages

```{r libraries} 
library(Seurat)
library(Signac)
library(reshape)
library(ggplot2)
library(GenomicRanges)
library(EnsDb.Hsapiens.v86)
```

## Parameters

```{r parameters}
cell_type = "PC"

# Paths
path_to_obj <- str_c(
  here::here("scATAC-seq/results/R_objects/level_4/"),
  cell_type,
    "/",
  cell_type,
  "_annotated_level_4.rds",
  sep = ""
)

path_to_save <- str_c(
  here::here("scATAC-seq/results/R_objects/level_4/"),
  cell_type,
    "/",
  cell_type,
  "_annotated_peak_calling_level_4.rds",
  sep = ""
)

color_palette<- c("black", "gray", "red", "yellow", "violet", 
                  "green4","blue", "chocolate1", "coral2", 
                  "blueviolet","brown1", "darkmagenta", 
                  "deepskyblue1", "dimgray", "deeppink1", 
                  "green", "lightgray", "hotpink1",
                  "indianred4", "khaki", "mediumorchid2")
```

## Functions

```{r functions}
peaks_quantification <- function(seurat_filtered, new_peaks){ 
    
  annotation <- GetGRangesFromEnsDb(ensdb = EnsDb.Hsapiens.v86,standard.chromosomes = T)
    seqlevelsStyle(annotation) <- "UCSC"
    genome(annotation) <- "hg38"

    macs2_counts <- FeatureMatrix(
    fragments = Fragments(seurat_filtered),
    features = new_peaks,
    cells = colnames(seurat_filtered))
    
    seurat_filtered[["peaks_redefined"]] <- CreateChromatinAssay(
      counts = macs2_counts, 
      genome = "hg38",
      fragments = Fragments(seurat_filtered),
      annotation = annotation)
    
    return(seurat_filtered)
}
```

# Load data

```{r}
seurat <- readRDS(path_to_obj)
```

Here, we can see the main clusters annotated thanks to the scRNA-seq data from the same cells and for the KNN algorithm.

```{r fig.wide=TRUE}
seurat <- SetIdent(seurat, 
                   value = seurat@meta.data$annotation_level_5)

table(seurat$annotation_level_5)
```


# New round of Peak Calling using MACS2 on each annotated cell type at level 1.

```{r}
peaks_level3 <- CallPeaks(
  object = seurat,
  group.by = "annotation_level_5",
  macs2.path = "/Users/pauli/opt/anaconda3/envs/Tonsil_atlas/bin/macs2")


#write.table(peaks_level3, path_to_save_peaks,
 #           quote=F, sep="\t", row.names=F, col.names=T)
```

```{r}
# remove peaks on nonstandard chromosomes and in genomic blacklist regions
peaks_level3 <- keepStandardChromosomes(peaks_level3, pruning.mode = "coarse")
peaks_level3 <- subsetByOverlaps(x = peaks_level3, 
                         ranges = blacklist_hg38_unified, 
                         invert = TRUE)
```

# Quantification of the new peak counts in the entire dataset

```{r}
seurat <- peaks_quantification(seurat, peaks_level3)
```

```{r}
# remove the initial peaks annotation
DefaultAssay(seurat) <- 'peaks_redefined'
seurat[['peaks_macs']] <- NULL
seurat

saveRDS(seurat,path_to_save)
```

# Session Information
```{r session}
sessionInfo()
```
